requirements.txt
fastapi
uvicorn
litellm==1.30.7
python-dotenv
httpx
rich
openai==1.12.0
pandas
